removing factorizations file
